+++ src/layouts/modules/theme-drawer/modules/general/modules/watermark-settings.vue
@@ -47,7 +47,7 @@ const isWatermarkTextVisible = computed(
         type="text"
         size="small"
         class="w-120px"
-        placeholder="SoybeanAdmin"
+        placeholder="DmpAdmin"
       />
     </SettingItem>
   </TransitionGroup>